Code review

+++ src/visual_package/Position.java
@@ -1,27 +1,27 @@
 package visual_package;
 
 public class Position {
-    private int xPos;
-    private int yPos;
+    private double xPos;
+    private double yPos;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.xPos = x;
         this.yPos = y;
     }
 
-    public int getxPos() {
+    public double getxPos() {
         return xPos;
     }
 
-    public void setxPos(int xPos) {
+    public void setxPos(double xPos) {
         this.xPos = xPos;
     }
 
-    public int getyPos() {
+    public double getyPos() {
         return yPos;
     }
 
-    public void setyPos(int yPos) {
+    public void setyPos(double yPos) {
         this.yPos = yPos;
     }
 }
